common/cpu_info_helpers: library to capture CPU information
ppc64_cpu tool does a lots of heavy lifting in finding the information
about the underlying CPU topology and various other details.  These
details might be required by other tools, probing the underlying
platform for similar details.  Moving such functions to a common
library, allows sharing across ppc64_cpu and others tools, instead of
multiple implementation of functions capturing same information.
